+++ Scripts/Data/Option.cs
@@ -1,0 +1,7 @@
+﻿namespace Sisk.SmarterSuit.Data {
+    public enum Option {
+        AlwaysAutoHelmet,
+        AdditionalFuelWarning,
+        FuelThreshold
+    }
+}