+++ Scripts/Data/Result.cs
@@ -1,0 +1,7 @@
+﻿namespace Sisk.SmarterSuit.Data {
+    public enum Result {
+        NoPermission,
+        Error,
+        Success
+    }
+}