+++ Scripts/Data/State.cs
@@ -1,4 +1,4 @@
-﻿namespace Sisk.SmarterSuit {
+﻿namespace Sisk.SmarterSuit.Data {
     /// <summary>
     ///     States to determine which actions should executed.
     /// </summary>
@@ -19,8 +19,8 @@
         Respawn,
 
         /// <summary>
-        ///     Extra step after a respawn if Pressurization is enabled
+        ///     Extra check for oxygen if Pressurization is enabled, because of false oxygen values for a couple of ticks.
         /// </summary>
-        CheckOxygenAfterRespawn
+        CheckOxygenAfterDelay
     }
 }