+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+<a name="1.2.0"></a>
+# [1.2.0](https://github.com/SiskSjet/SmartSuit/compare/v1.1.2...v1.2.0) (2018-09-23)
+
+
+### Bug Fixes
+
+* fix multiple bug in multiplayer. ([4da4a5](https://github.com/SiskSjet/SmartSuit/commit/4da4a5))
+
+### Features
+
+* add an optional feature for to always check if helmet is needed ([ba1aa38](https://github.com/SiskSjet/SmartSuit/commit/ba1aa38))
+* add an optional feature to enable an additional fuel warning ([fe533dc](https://github.com/SiskSjet/SmartSuit/commit/fe533dc))
+* set speed only if jetpack is enabled ([17aa04e](https://github.com/SiskSjet/SmartSuit/commit/17aa04e))
+
+
+
 <a name="1.1.2"></a>
 # [1.1.2](https://github.com/SiskSjet/SmartSuit/compare/v1.1.1...v1.1.2) (2018-09-21)
 
@@ -20,7 +36,7 @@
 
 
 <a name="1.1.0"></a>
-# [1.1.0](https://github.com/SiskSjet/SmartSuit/compare/v1.0.1...v1.1.0) (2018-09-19)
+# [1.1.0](https://github.com/SiskSjet/SmartSuit/compare/v1.0.2...v1.1.0) (2018-09-19)
 
 
 ### Bug Fixes
@@ -62,6 +78,23 @@
 ### Features
 
 * add initial features ([4cdf40](https://github.com/SiskSjet/SmartSuit/commit/4cdf40))
+
+
+
+<a name=""></a>
+#  (2018-09-19)
+
+
+### Bug Fixes
+
+* fix a possible game crash on new created worlds ([75a6155](https://github.com/SiskSjet/SmartSuit/commit/75a6155))
+* fix a prossible NRE when spaning as space suit in a new world ([70bab3b](https://github.com/SiskSjet/SmartSuit/commit/70bab3b))
+
+
+### Features
+
+* add support for the 'Remove all automatic jetpack activation' mod ([f87d991](https://github.com/SiskSjet/SmartSuit/commit/f87d991))
+* improved dampener handling in gravity with ground in range ([9ae97e6](https://github.com/SiskSjet/SmartSuit/commit/9ae97e6))